speedsensor_app/requirements.txt
PySide6>=6.6.0
pyserial>=3.5
pyqtgraph>=0.13.0
numpy>=1.24.0
openpyxl>=3.1.0
